fix: Remove workaround in demo example for bevy_gizmo (#56)
Just a small tweak to the highlight_nearest_sphere function in
examples/demo.rs. I swapped out the old way of pulling out translation
and scale from the transform and replaced it with the more
straightforward .translation() and .scale() calls. It’s cleaner, easier
to read, and allows gizmo to handling rotation.

Tested (ran the demo) on linux under wayland with bevy 0.16.1 and
nightly rust

Co-authored-by: Aevyrie <aevyrie@gmail.com>